+++ R/plot_perturb.R
@@ -187,9 +187,7 @@ plot_perturb <- function(input,
     raster::crs(lrr_mean) <- sp::CRS(SRS_string = cref0)
     if (!is.null(cref1)) {
       lrr_mean <- raster::projectRaster(lrr_mean,
-                                        crs = sp::CRS(SRS_string = cref1),
-                                        method = "ngb",
-                                        legacy = TRUE)
+                                        crs = sp::CRS(SRS_string = cref1))
     }
     
     lrr_sd <- spatstat.geom::pixellate(input$predict,
@@ -198,9 +196,7 @@ plot_perturb <- function(input,
     raster::crs(lrr_sd) <- sp::CRS(SRS_string = cref0)
     if (!is.null(cref1)) {
       lrr_sd <- raster::projectRaster(lrr_sd,
-                                      crs = sp::CRS(SRS_string = cref1),
-                                      method = "ngb",
-                                      legacy = TRUE)
+                                      crs = sp::CRS(SRS_string = cref1))
     }
     
     pval_mean <- spatstat.geom::pixellate(input$predict,
@@ -209,9 +205,7 @@ plot_perturb <- function(input,
     raster::crs(pval_mean) <- sp::CRS(SRS_string = cref0)
     if (!is.null(cref1)) {
       pval_mean <- raster::projectRaster(pval_mean,
-                                         crs = sp::CRS(SRS_string = cref1),
-                                         method = "ngb",
-                                         legacy = TRUE)
+                                         crs = sp::CRS(SRS_string = cref1))
     }
     
     pval_prop <- spatstat.geom::pixellate(input$predict,
@@ -220,9 +214,7 @@ plot_perturb <- function(input,
     raster::crs(pval_prop) <- sp::CRS(SRS_string = cref0)
     if (!is.null(cref1)) {
       pval_prop <- raster::projectRaster(pval_prop,
-                                         crs = sp::CRS(SRS_string = cref1),
-                                         method = "ngb",
-                                         legacy = TRUE)
+                                         crs = sp::CRS(SRS_string = cref1))
     }
     
     # Plot 5: mean log relative risk